+++ homematicip/aio/group.py
@@ -3,6 +3,8 @@ import json
 from homematicip.base.enums import *
 from homematicip.group import (
     AlarmSwitchingGroup,
+    AccessAuthorizationProfileGroup,
+    AccessControlGroup,
     EnvironmentGroup,
     ExtendedLinkedShutterGroup,
     ExtendedLinkedSwitchingGroup,
@@ -21,6 +23,7 @@ from homematicip.group import (
     HotWaterGroup,
     HumidityWarningRuleGroup,
     InboxGroup,
+    IndoorClimateGroup,
     LinkedSwitchingGroup,
     LockOutProtectionRule,
     MetaGroup,
@@ -185,6 +188,10 @@ class AsyncInboxGroup(InboxGroup, AsyncGroup):
     pass
 
 
+class AsyncIndoorClimateGroup(IndoorClimateGroup, AsyncGroup):
+    pass
+
+
 class AsyncSecurityZoneGroup(SecurityZoneGroup, AsyncGroup):
     pass
 
@@ -276,3 +283,11 @@ class AsyncEnvironmentGroup(EnvironmentGroup, AsyncGroup):
 class AsyncHotWaterGroup(HotWaterGroup, AsyncGroup):
     async def set_profile_mode(self, profileMode: ProfileMode):
         return await self._connection.api_call(*super().set_profile_mode(profileMode))
+
+
+class AsyncAccessAuthorizationProfileGroup(AccessAuthorizationProfileGroup, AsyncGroup):
+    pass
+
+
+class AsyncAccessControlGroup(AccessControlGroup, AsyncGroup):
+    pass