+++ homematicip/base/functionalChannels.py
@@ -6,7 +6,7 @@ from homematicip.group import Group
 
 
 class FunctionalChannel(HomeMaticIPObject):
-    """ this is the base class for the functional channels """
+    """this is the base class for the functional channels"""
 
     def __init__(self):
         super().__init__(None)
@@ -21,9 +21,9 @@ class FunctionalChannel(HomeMaticIPObject):
         self._connection = None
 
     def from_json(self, js, groups: Iterable[Group]):
-        """ this function will load the functional channel object 
-        from a json object and the given groups 
-        
+        """this function will load the functional channel object
+        from a json object and the given groups
+
         Args:
             js(dict): the json object
             groups(Iterable[Group]): the groups for referencing
@@ -43,9 +43,16 @@ class FunctionalChannel(HomeMaticIPObject):
 
         super().from_json(js)
 
+    def __str__(self):
+        return "{} {} Index({})".format(
+            self.functionalChannelType,
+            self.label if self.label else "unknown",
+            self.index,
+        )
+
 
 class DeviceBaseChannel(FunctionalChannel):
-    """ this is the representative of the DEVICE_BASE channel"""
+    """this is the representative of the DEVICE_BASE channel"""
 
     def __init__(self):
         super().__init__()
@@ -94,7 +101,7 @@ class DeviceBaseChannel(FunctionalChannel):
 
 
 class AccessControllerChannel(DeviceBaseChannel):
-    """ this is the representative of the ACCESS_CONTROLLER_CHANNEL channel"""
+    """this is the representative of the ACCESS_CONTROLLER_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -110,7 +117,7 @@ class AccessControllerChannel(DeviceBaseChannel):
 
 
 class DeviceSabotageChannel(DeviceBaseChannel):
-    """ this is the representative of the DEVICE_SABOTAGE channel"""
+    """this is the representative of the DEVICE_SABOTAGE channel"""
 
     def __init__(self):
         super().__init__()
@@ -122,7 +129,7 @@ class DeviceSabotageChannel(DeviceBaseChannel):
 
 
 class DeviceOperationLockChannel(DeviceBaseChannel):
-    """ this is the representative of the DEVICE_OPERATIONLOCK channel"""
+    """this is the representative of the DEVICE_OPERATIONLOCK channel"""
 
     def __init__(self):
         super().__init__()
@@ -134,7 +141,7 @@ class DeviceOperationLockChannel(DeviceBaseChannel):
 
 
 class DeviceIncorrectPositionedChannel(DeviceBaseChannel):
-    """ this is the representative of the DEVICE_INCORRECT_POSITIONED channel"""
+    """this is the representative of the DEVICE_INCORRECT_POSITIONED channel"""
 
     def __init__(self):
         super().__init__()
@@ -146,7 +153,7 @@ class DeviceIncorrectPositionedChannel(DeviceBaseChannel):
 
 
 class DevicePermanentFullRxChannel(DeviceBaseChannel):
-    """ this is the representative of the DEVICE_PERMANENT_FULL_RX channel"""
+    """this is the representative of the DEVICE_PERMANENT_FULL_RX channel"""
 
     def __init__(self):
         super().__init__()
@@ -157,8 +164,20 @@ class DevicePermanentFullRxChannel(DeviceBaseChannel):
         self.permanentFullRx = js["permanentFullRx"]
 
 
+class AccessAuthorizationChannel(FunctionalChannel):
+    """this represents ACCESS_AUTHORIZATION_CHANNEL channel"""
+
+    def __init__(self):
+        super().__init__()
+        self.authorized = False
+
+    def from_json(self, js, groups: Iterable[Group]):
+        super().from_json(js, groups)
+        self.authorized = js["authorized"]
+
+
 class WaterSensorChannel(FunctionalChannel):
-    """ this is the representative of the WATER_SENSOR_CHANNEL channel"""
+    """this is the representative of the WATER_SENSOR_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -192,7 +211,7 @@ class WaterSensorChannel(FunctionalChannel):
 
 
 class HeatingThermostatChannel(FunctionalChannel):
-    """ this is the representative of the HEATING_THERMOSTAT_CHANNEL channel"""
+    """this is the representative of the HEATING_THERMOSTAT_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -219,7 +238,7 @@ class HeatingThermostatChannel(FunctionalChannel):
 
 
 class ShutterContactChannel(FunctionalChannel):
-    """ this is the representative of the SHUTTER_CONTACT_CHANNEL channel"""
+    """this is the representative of the SHUTTER_CONTACT_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -233,11 +252,11 @@ class ShutterContactChannel(FunctionalChannel):
 
 
 class RotaryHandleChannel(ShutterContactChannel):
-    """ this is the representative of the ROTARY_HANDLE_CHANNEL channel"""
+    """this is the representative of the ROTARY_HANDLE_CHANNEL channel"""
 
 
 class ContactInterfaceChannel(ShutterContactChannel):
-    """ this is the representative of the CONTACT_INTERFACE_CHANNEL channel"""
+    """this is the representative of the CONTACT_INTERFACE_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -251,7 +270,7 @@ class ContactInterfaceChannel(ShutterContactChannel):
 
 
 class ClimateSensorChannel(FunctionalChannel):
-    """ this is the representative of the CLIMATE_SENSOR_CHANNEL channel"""
+    """this is the representative of the CLIMATE_SENSOR_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -267,7 +286,7 @@ class ClimateSensorChannel(FunctionalChannel):
 
 
 class DoorChannel(FunctionalChannel):
-    """ this is the representative of the DoorChannel channel"""
+    """this is the representative of the DoorChannel channel"""
 
     def __init__(self):
         super().__init__()
@@ -284,8 +303,51 @@ class DoorChannel(FunctionalChannel):
         self.ventilationPositionSupported = js["ventilationPositionSupported"]
 
 
+class DoorLockChannel(FunctionalChannel):
+    """This respresents of the DoorLockChannel"""
+
+    def __init__(self):
+        super().__init__()
+        self.autoRelockDelay = False
+        self.doorHandleType = "UNKNOWN"
+        self.doorLockDirection = False
+        self.doorLockNeutralPosition = False
+        self.doorLockTurns = False
+        self.lockState = LockState.UNLOCKED
+        self.motorState = MotorState.STOPPED
+
+    def from_json(self, js, groups: Iterable[Group]):
+        super().from_json(js, groups)
+
+        self.autoRelockDelay = js["autoRelockDelay"]
+        self.doorHandleType = js["doorHandleType"]
+        self.doorLockDirection = js["doorLockDirection"]
+        self.doorLockNeutralPosition = js["doorLockNeutralPosition"]
+        self.doorLockTurns = js["doorLockTurns"]
+        self.lockState = LockState.from_str(js["lockState"])
+        self.motorState = MotorState.from_str(js["motorState"])
+
+
+class DoorLockSensorChannel(FunctionalChannel):
+    """This respresents of the DoorLockSensorChannel"""
+
+    def __init__(self):
+        super().__init__()
+        self.doorLockDirection = False
+        self.doorLockNeutralPosition = False
+        self.doorLockTurns = False
+        self.lockState = LockState.UNLOCKED
+
+    def from_json(self, js, groups: Iterable[Group]):
+        super().from_json(js, groups)
+        self.doorLockDirection = js["doorLockDirection"]
+        self.doorLockNeutralPosition = js["doorLockNeutralPosition"]
+        self.doorLockTurns = js["doorLockTurns"]
+        self.lockState = LockState.from_str(js["lockState"])
+
+
 class WallMountedThermostatWithoutDisplayChannel(ClimateSensorChannel):
-    """ this is the representative of the WALL_MOUNTED_THERMOSTAT_WITHOUT_DISPLAY_CHANNEL channel"""
+    """this is the representative of the WALL_MOUNTED_THERMOSTAT_WITHOUT_DISPLAY_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -297,7 +359,7 @@ class WallMountedThermostatWithoutDisplayChannel(ClimateSensorChannel):
 
 
 class AnalogRoomControlChannel(FunctionalChannel):
-    """ this is the representative of the ANALOG_ROOM_CONTROL_CHANNEL channel"""
+    """this is the representative of the ANALOG_ROOM_CONTROL_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -313,7 +375,7 @@ class AnalogRoomControlChannel(FunctionalChannel):
 
 
 class WallMountedThermostatProChannel(WallMountedThermostatWithoutDisplayChannel):
-    """ this is the representative of the WALL_MOUNTED_THERMOSTAT_PRO_CHANNEL channel"""
+    """this is the representative of the WALL_MOUNTED_THERMOSTAT_PRO_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -327,7 +389,7 @@ class WallMountedThermostatProChannel(WallMountedThermostatWithoutDisplayChannel
 
 
 class SmokeDetectorChannel(FunctionalChannel):
-    """ this is the representative of the SMOKE_DETECTOR_CHANNEL channel"""
+    """this is the representative of the SMOKE_DETECTOR_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -341,7 +403,7 @@ class SmokeDetectorChannel(FunctionalChannel):
 
 
 class SwitchChannel(FunctionalChannel):
-    """ this is the representative of the SWITCH_CHANNEL channel"""
+    """this is the representative of the SWITCH_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -357,7 +419,7 @@ class SwitchChannel(FunctionalChannel):
 
 
 class SwitchMeasuringChannel(SwitchChannel):
-    """ this is the representative of the SWITCH_MEASURING_CHANNEL channel"""
+    """this is the representative of the SWITCH_MEASURING_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -371,7 +433,7 @@ class SwitchMeasuringChannel(SwitchChannel):
 
 
 class DeviceGlobalPumpControlChannel(DeviceBaseChannel):
-    """ this is the representative of the DEVICE_GLOBAL_PUMP_CONTROL channel"""
+    """this is the representative of the DEVICE_GLOBAL_PUMP_CONTROL channel"""
 
     def __init__(self):
         super().__init__()
@@ -398,7 +460,7 @@ class DeviceGlobalPumpControlChannel(DeviceBaseChannel):
 
 
 class DeviceBaseFloorHeatingChannel(DeviceBaseChannel):
-    """ this is the representative of the DEVICE_BASE_FLOOR_HEATING channel"""
+    """this is the representative of the DEVICE_BASE_FLOOR_HEATING channel"""
 
     def __init__(self):
         super().__init__()
@@ -416,7 +478,7 @@ class DeviceBaseFloorHeatingChannel(DeviceBaseChannel):
 
 
 class MotionDetectionChannel(FunctionalChannel):
-    """ this is the representative of the MOTION_DETECTION_CHANNEL channel"""
+    """this is the representative of the MOTION_DETECTION_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -442,7 +504,7 @@ class MotionDetectionChannel(FunctionalChannel):
 
 
 class PresenceDetectionChannel(FunctionalChannel):
-    """ this is the representative of the PRESENCE_DETECTION_CHANNEL channel"""
+    """this is the representative of the PRESENCE_DETECTION_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -466,7 +528,7 @@ class PresenceDetectionChannel(FunctionalChannel):
 
 
 class ShutterChannel(FunctionalChannel):
-    """ this is the representative of the SHUTTER_CHANNEL channel"""
+    """this is the representative of the SHUTTER_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -506,7 +568,7 @@ class ShutterChannel(FunctionalChannel):
 
 
 class BlindChannel(ShutterChannel):
-    """ this is the representative of the BLIND_CHANNEL channel"""
+    """this is the representative of the BLIND_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -525,7 +587,7 @@ class BlindChannel(ShutterChannel):
 
 
 class MultiModeInputBlindChannel(BlindChannel):
-    """ this is the representative of the MULTI_MODE_INPUT_BLIND_CHANNEL channel"""
+    """this is the representative of the MULTI_MODE_INPUT_BLIND_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -544,7 +606,7 @@ class MultiModeInputBlindChannel(BlindChannel):
 
 
 class DimmerChannel(FunctionalChannel):
-    """ this is the representative of the DIMMER_CHANNEL channel"""
+    """this is the representative of the DIMMER_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -560,7 +622,7 @@ class DimmerChannel(FunctionalChannel):
 
 
 class WeatherSensorChannel(FunctionalChannel):
-    """ this is the representative of the WEATHER_SENSOR_CHANNEL channel"""
+    """this is the representative of the WEATHER_SENSOR_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -594,7 +656,7 @@ class WeatherSensorChannel(FunctionalChannel):
 
 
 class WeatherSensorPlusChannel(WeatherSensorChannel):
-    """ this is the representative of the WEATHER_SENSOR_PLUS_CHANNEL channel"""
+    """this is the representative of the WEATHER_SENSOR_PLUS_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -612,7 +674,7 @@ class WeatherSensorPlusChannel(WeatherSensorChannel):
 
 
 class WeatherSensorProChannel(WeatherSensorPlusChannel):
-    """ this is the representative of the WEATHER_SENSOR_PRO_CHANNEL channel"""
+    """this is the representative of the WEATHER_SENSOR_PRO_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -628,19 +690,19 @@ class WeatherSensorProChannel(WeatherSensorPlusChannel):
 
 
 class SingleKeyChannel(FunctionalChannel):
-    """ this is the representative of the SINGLE_KEY_CHANNEL channel"""
+    """this is the representative of the SINGLE_KEY_CHANNEL channel"""
 
 
 class AlarmSirenChannel(FunctionalChannel):
-    """ this is the representative of the ALARM_SIREN_CHANNEL channel"""
+    """this is the representative of the ALARM_SIREN_CHANNEL channel"""
 
 
 class FloorTeminalBlockChannel(FunctionalChannel):
-    """ this is the representative of the FLOOR_TERMINAL_BLOCK_CHANNEL channel"""
+    """this is the representative of the FLOOR_TERMINAL_BLOCK_CHANNEL channel"""
 
 
 class FloorTerminalBlockLocalPumpChannel(FunctionalChannel):
-    """ this is the representative of the FLOOR_TERMINAL_BLOCK_LOCAL_PUMP_CHANNEL  channel"""
+    """this is the representative of the FLOOR_TERMINAL_BLOCK_LOCAL_PUMP_CHANNEL  channel"""
 
     def __init__(self):
         super().__init__()
@@ -658,15 +720,15 @@ class FloorTerminalBlockLocalPumpChannel(FunctionalChannel):
 
 
 class HeatDemandChannel(FunctionalChannel):
-    """ this is the representative of the HEAT_DEMAND_CHANNEL channel"""
+    """this is the representative of the HEAT_DEMAND_CHANNEL channel"""
 
 
 class DehumidifierDemandChannel(FunctionalChannel):
-    """ this is the representative of the DEHUMIDIFIER_DEMAND_CHANNEL channel"""
+    """this is the representative of the DEHUMIDIFIER_DEMAND_CHANNEL channel"""
 
 
 class PassageDetectorChannel(FunctionalChannel):
-    """ this is the representative of the PASSAGE_DETECTOR_CHANNEL channel"""
+    """this is the representative of the PASSAGE_DETECTOR_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -690,7 +752,7 @@ class PassageDetectorChannel(FunctionalChannel):
 
 
 class InternalSwitchChannel(FunctionalChannel):
-    """ this is the representative of the INTERNAL_SWITCH_CHANNEL channel"""
+    """this is the representative of the INTERNAL_SWITCH_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -708,12 +770,13 @@ class InternalSwitchChannel(FunctionalChannel):
         self.valveProtectionDuration = js["valveProtectionDuration"]
         self.valveProtectionSwitchingInterval = js["valveProtectionSwitchingInterval"]
 
+
 class ImpulseOutputChannel(FunctionalChannel):
-    """ this is the representation of the IMPULSE_OUTPUT_CHANNEL """
+    """this is the representation of the IMPULSE_OUTPUT_CHANNEL"""
 
     def __init__(self):
         super().__init__()
-        
+
     def from_json(self, js, groups: Iterable[Group]):
         super().from_json(js, groups)
         self.impulseDuration = js["impulseDuration"]
@@ -721,7 +784,7 @@ class ImpulseOutputChannel(FunctionalChannel):
 
 
 class MultiModeInputChannel(FunctionalChannel):
-    """ this is the representative of the MULTI_MODE_INPUT_CHANNEL channel"""
+    """this is the representative of the MULTI_MODE_INPUT_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -737,7 +800,7 @@ class MultiModeInputChannel(FunctionalChannel):
 
 
 class NotificationLightChannel(DimmerChannel):
-    """ this is the representative of the NOTIFICATION_LIGHT_CHANNEL channel"""
+    """this is the representative of the NOTIFICATION_LIGHT_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -753,7 +816,7 @@ class NotificationLightChannel(DimmerChannel):
 
 
 class LightSensorChannel(FunctionalChannel):
-    """ this is the representative of the LIGHT_SENSOR_CHANNEL channel"""
+    """this is the representative of the LIGHT_SENSOR_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -775,11 +838,11 @@ class LightSensorChannel(FunctionalChannel):
 
 
 class GenericInputChannel(FunctionalChannel):
-    """ this is the representative of the GENERIC_INPUT_CHANNEL channel"""
+    """this is the representative of the GENERIC_INPUT_CHANNEL channel"""
 
 
 class AnalogOutputChannel(FunctionalChannel):
-    """ this is the representative of the ANALOG_OUTPUT_CHANNEL channel"""
+    """this is the representative of the ANALOG_OUTPUT_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -792,7 +855,7 @@ class AnalogOutputChannel(FunctionalChannel):
 
 
 class AccelerationSensorChannel(FunctionalChannel):
-    """ this is the representative of the ACCELERATION_SENSOR_CHANNEL channel"""
+    """this is the representative of the ACCELERATION_SENSOR_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -838,7 +901,7 @@ class AccelerationSensorChannel(FunctionalChannel):
 
 
 class DeviceRechargeableWithSabotage(DeviceSabotageChannel):
-    """ this is the representative of the DEVICE_RECHARGEABLE_WITH_SABOTAGE channel"""
+    """this is the representative of the DEVICE_RECHARGEABLE_WITH_SABOTAGE channel"""
 
     def __init__(self):
         super().__init__()
@@ -851,24 +914,26 @@ class DeviceRechargeableWithSabotage(DeviceSabotageChannel):
 
 
 class FloorTerminalBlockMechanicChannel(FunctionalChannel):
-    """ this is the representative of the class FLOOR_TERMINAL_BLOCK_MECHANIC_CHANNEL(FunctionalChannel) channel"""
+    """this is the representative of the class FLOOR_TERMINAL_BLOCK_MECHANIC_CHANNEL(FunctionalChannel) channel"""
 
     def __init__(self):
         super().__init__()
         #:ValveState:the current valve state
         self.valveState = ValveState.ADAPTION_DONE
+        self.valvePosition = 0.0
 
     def from_json(self, js, groups: Iterable[Group]):
         super().from_json(js, groups)
         self.set_attr_from_dict("valveState", js)
+        self.set_attr_from_dict("valvePosition", js)
 
 
 class ChangeOverChannel(FunctionalChannel):
-    """ this is the representative of the CHANGE_OVER_CHANNEL channel"""
+    """this is the representative of the CHANGE_OVER_CHANNEL channel"""
 
 
 class MainsFailureChannel(FunctionalChannel):
-    """ this is the representative of the MAINS_FAILURE_CHANNEL channel"""
+    """this is the representative of the MAINS_FAILURE_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -882,7 +947,7 @@ class MainsFailureChannel(FunctionalChannel):
 
 
 class MultiModeInputSwitchChannel(FunctionalChannel):
-    """ this is the representative of the MULTI_MODE_INPUT_SWITCH_CHANNEL channel"""
+    """this is the representative of the MULTI_MODE_INPUT_SWITCH_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -902,7 +967,7 @@ class MultiModeInputSwitchChannel(FunctionalChannel):
 
 
 class MultiModeInputDimmerChannel(FunctionalChannel):
-    """ this is the representative of the MULTI_MODE_INPUT_DIMMER_CHANNEL channel  """
+    """this is the representative of the MULTI_MODE_INPUT_DIMMER_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -924,7 +989,7 @@ class MultiModeInputDimmerChannel(FunctionalChannel):
 
 
 class TiltVibrationSensorChannel(FunctionalChannel):
-    """ this is the representative of the TILT_VIBRATION_SENSOR_CHANNEL channel"""
+    """this is the representative of the TILT_VIBRATION_SENSOR_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -953,7 +1018,7 @@ class TiltVibrationSensorChannel(FunctionalChannel):
 
 
 class ShadingChannel(FunctionalChannel):
-    """ this is the representative of the SHADING_CHANNEL channel"""
+    """this is the representative of the SHADING_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -1016,7 +1081,7 @@ class ShadingChannel(FunctionalChannel):
 
 
 class RainDetectionChannel(FunctionalChannel):
-    """ this is the representative of the TILT_VIBRATION_SENSOR_CHANNEL channel"""
+    """this is the representative of the TILT_VIBRATION_SENSOR_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -1030,8 +1095,9 @@ class RainDetectionChannel(FunctionalChannel):
         self.set_attr_from_dict("rainSensorSensitivity", js)
         self.set_attr_from_dict("raining", js)
 
+
 class TemperatureDifferenceSensor2Channel(FunctionalChannel):
-    """ this is the representative of the TEMPERATURE_SENSOR_2_EXTERNAL_DELTA_CHANNEL channel """
+    """this is the representative of the TEMPERATURE_SENSOR_2_EXTERNAL_DELTA_CHANNEL channel"""
 
     def __init__(self):
         super().__init__()
@@ -1041,6 +1107,7 @@ class TemperatureDifferenceSensor2Channel(FunctionalChannel):
         self.temperatureExternalOne = 0.0
         #:float:
         self.temperatureExternalTwo = 0.0
+
     def from_json(self, js, groups: Iterable[Group]):
         super().from_json(js, groups)
         self.set_attr_from_dict("temperatureExternalDelta", js)