+++ tests/aio_tests/test_async_groups.py
@@ -348,3 +348,41 @@ async def test_switching_alarm_group(no_ssl_fake_async_home: AsyncHome):
     assert g.signalAcoustic == AcousticAlarmSignal.FREQUENCY_HIGHON_OFF
     assert g.signalOptical == OpticalAlarmSignal.BLINKING_ALTERNATELY_REPEATING
     assert g.onTime == 5
+
+
+@pytest.mark.asyncio
+async def test_access_control(no_ssl_fake_async_home: AsyncHome):
+    g = no_ssl_fake_async_home.search_group_by_id(
+        "00000000-0000-0000-0000-000000000033"
+    )
+    await no_ssl_fake_async_home.get_current_state()
+
+    assert str(g) == "ACCESS_CONTROL AmHaustuere2"
+
+
+@pytest.mark.asyncio
+async def test_access_authorization_profile_group(no_ssl_fake_async_home: AsyncHome):
+    g = no_ssl_fake_async_home.search_group_by_id(
+        "00000000-0000-0000-0000-000000000032"
+    )
+    await no_ssl_fake_async_home.get_current_state()
+
+    assert g.label == "Walter"
+    assert g.active == True
+    assert g.authorizationPinAssigned == True
+    assert g.authorized == True
+
+
+@pytest.mark.asyncio
+async def test_indoor_climate_group(no_ssl_fake_async_home: AsyncHome):
+    g = no_ssl_fake_async_home.search_group_by_id(
+        "00000000-0000-0000-0000-0000000000IC"
+    )
+    await no_ssl_fake_async_home.get_current_state()
+
+    assert g.label == "Stanovanje"
+    assert g.sabotage == False
+    assert g.ventilationLevel == 0.5
+    assert g.ventilationState == None
+    assert g.windowState == WindowState.CLOSED
+    assert str(g) == "INDOOR_CLIMATE Stanovanje sabotage(False) ventilationLevel(0.5) ventilationState(None) windowState(CLOSED)"